requirements.txt
sentence-transformers
streamlit
torch
farm-haystack

